プリセットを1ファイル=1関数に分割して tree-shaking を改善
- presetRules.ts, presetKeyboardGuide.ts, presetKeyboardLayout.ts を
  src/impl/presets/ 配下の個別ファイルに分割
- barrel re-export (presets/index.ts) で公開 API を維持
- package.json に subpath exports (./presets/*) を追加
- テストファイルを presets/ 配下に移動
