+++ package.json
@@ -18,9 +18,17 @@
   "module": "./dist/index.js",
   "sideEffects": false,
   "exports": {
-    "import": {
-      "types": "./dist/types/index.d.ts",
-      "default": "./dist/index.js"
+    ".": {
+      "import": {
+        "types": "./dist/types/index.d.ts",
+        "default": "./dist/index.js"
+      }
+    },
+    "./presets/*": {
+      "import": {
+        "types": "./dist/types/impl/presets/*.d.ts",
+        "default": "./dist/impl/presets/*.js"
+      }
     }
   },
   "scripts": {
