+++ src/impl/presets/keyboardGuideJis106Nicola.ts
@@ -1,0 +1,6 @@
+import jis_106_nicola from "../../assets/keyboard_guides/jis_106_nicola.json";
+import { loadJsonKeyboardGuide } from "../keyboardGuideLoader";
+
+export function loadPresetKeyboardGuideJis106Nicola() {
+  return loadJsonKeyboardGuide(jis_106_nicola);
+}
