+++ src/impl/presets/keyboardGuideUs101Default.ts
@@ -1,0 +1,6 @@
+import us_101_default from "../../assets/keyboard_guides/us_101_default.json";
+import { loadJsonKeyboardGuide } from "../keyboardGuideLoader";
+
+export function loadPresetKeyboardGuideUs101Default() {
+  return loadJsonKeyboardGuide(us_101_default);
+}
