+++ src/impl/presets/keyboardGuideUsHhkbDefault.ts
@@ -1,0 +1,6 @@
+import us_hhkb_default from "../../assets/keyboard_guides/us_hhkb_default.json";
+import { loadJsonKeyboardGuide } from "../keyboardGuideLoader";
+
+export function loadPresetKeyboardGuideUsHhkbDefault() {
+  return loadJsonKeyboardGuide(us_hhkb_default);
+}
