+++ src/impl/presets/keyboardLayoutDvorak.ts
@@ -1,0 +1,6 @@
+import dvorak from "../../assets/keyboard_layouts/dvorak.json";
+import { loadJsonKeyboardLayout } from "../keyboardLayoutLoader";
+
+export function loadPresetKeyboardLayoutDvorak() {
+  return loadJsonKeyboardLayout(dvorak);
+}
