+++ src/index.ts
@@ -21,15 +21,13 @@ export { KeyboardGuide, type KeyRect, type KeyTop } from "./impl/keyboardGuide";
 export { detectKeyboardLayout } from "./browser/osKeyboardLayout";
 export * from "./impl/alphaNumericRule";
 export { loadJsonKeyboardLayout } from "./impl/keyboardLayoutLoader";
-export * from "./impl/presetKeyboardLayout";
 
 export { loadJsonRule } from "./impl/jsonRuleLoader";
 export { loadMozcRule } from "./impl/mozcRuleLoader";
 
 export { loadJsonKeyboardGuide } from "./impl/keyboardGuideLoader";
 export type { PhysicalKeyboardLayoutName } from "./impl/keyboardGuide";
-export * from "./impl/presetKeyboardGuide";
-export * from "./impl/presetRules";
+export * from "./impl/presets";
 export * from "./impl/stats";
 
 export { build } from "./core/automaton";
